requirements.txt
flask
requests
python-dotenv
fpdf
python-pptx
